Added support for DEFAULT in COPY FROM
Previous to this commit, COPY FROM command used to load the column
DEFAULT value only if the column was missing in the command specification.

With this commit we introduce a new feature that works like the NULL
feature from COPY command. The user will be able to specify a marker,
and whenever that marker is found in the input of COPY FROM, it will
be replaced with the DEFAULT value of the corresponding column.

In order to make "the new" COPY from backward compatible, the DEFAULT
feature will only take place if the use specifies some value for that
option.

We are taking advantage of the code that was already implemented in
COPY FROM to find and evaluate the DEFAULT expressions, with the only
difference that now we check both for columns that are missing in the
command specification, and for columns where the DEFAULT marker was
found.

This commit also changes the `file_fdw` to make it possible to use
the new DEFAULT option of COPY FROM. We had to introduce the tuple
context in order to be able to evaluate the DEFAULT value in `file_fdw`.

Signed-off-by: Israel Barth <barthisrael@gmail.com>